Framework (#161)
* session inherit from ABC

* add module interface
