+++ tatk/nlu/nlu.py
@@ -1,8 +1,9 @@
 """Natural language understanding interface."""
-from abc import ABC, abstractmethod
+from abc import abstractmethod
+from tatk.util.module import Module
 
 
-class NLU(ABC):
+class NLU(Module):
     """NLU module interface."""
 
     def predict(self, utterance, context=list()):
@@ -37,23 +38,5 @@ class NLU(ABC):
         return [[]]
 
     @abstractmethod
-    def train(self, *args, **kwargs):
-        """Model training entry point"""
-        pass
-
-    @abstractmethod
-    def test(self, *args, **kwargs):
-        """Model testing entry point"""
-        pass
-
-    def from_cache(self, *args, **kwargs):
-        """restore internal state for multi-turn dialog"""
-        return None
-
-    def to_cache(self, *args, **kwargs):
-        """save internal state for multi-turn dialog"""
-        return None
-
-    def init_session(self):
-        """Init the class variables for a new session."""
-        pass
+    def get_name(self):
+        return super(NLU, self).get_name() + '-' + 'NLU'
