+++ tatk/policy/policy.py
@@ -1,8 +1,9 @@
 """Policy Interface"""
-from abc import ABC, abstractmethod
+from abc import abstractmethod
+from tatk.util.module import Module
 
 
-class Policy(ABC):
+class Policy(Module):
     """Base class for policy model."""
 
     def predict(self, state):
@@ -33,23 +34,5 @@ class Policy(ABC):
         pass
 
     @abstractmethod
-    def train(self, *args, **kwargs):
-        """Model training entry point"""
-        pass
-
-    @abstractmethod
-    def test(self, *args, **kwargs):
-        """Model testing entry point"""
-        pass
-
-    def from_cache(self, *args, **kwargs):
-        """restore internal state for multi-turn dialog"""
-        return None
-
-    def to_cache(self, *args, **kwargs):
-        """save internal state for multi-turn dialog"""
-        return None
-
-    def init_session(self):
-        """Init the class variables for a new session."""
-        pass
+    def get_name(self):
+        return super(Policy, self).get_name() + '-' + 'Policy'
