+++ tatk/util/module.py
@@ -1,0 +1,31 @@
+"""Natural language understanding interface."""
+from abc import ABC, abstractmethod
+
+
+class Module(ABC):
+
+    @abstractmethod
+    def train(self, *args, **kwargs):
+        """Model training entry point"""
+        pass
+
+    @abstractmethod
+    def test(self, *args, **kwargs):
+        """Model testing entry point"""
+        pass
+
+    @abstractmethod
+    def get_name(self):
+        return "Module"
+
+    def from_cache(self, *args, **kwargs):
+        """restore internal state for multi-turn dialog"""
+        return None
+
+    def to_cache(self, *args, **kwargs):
+        """save internal state for multi-turn dialog"""
+        return None
+
+    def init_session(self):
+        """Init the class variables for a new session."""
+        pass
